Ajustes finales para analisis

+++ sx-compras-server/grails-app/domain/sx/core/Proveedor.groovy
@@ -37,7 +37,7 @@ class Proveedor {
 
     Long sw2
 
-    Long descuentoF = 0.0
+    BigDecimal descuentoF = 0.0
 
     Long diasDF = 0
